serving/requirements.txt
gunicorn
label-studio-ml
numpy<1.24,>=1.22
# psycopg2-binary
pillow
rq
minio

# https://developer.apple.com/metal/tensorflow-plugin/
tensorflow~=2.12.0; sys_platform != 'darwin'
tensorflow-macos~=2.12.0; sys_platform == 'darwin'
tensorflow-metal~=0.8.0; sys_platform == 'darwin'